ganti migration down pada finance_mutation_schedules

+++ database/migrations/2021_09_08_074959_alter_table_finance_mutation_schedules_modify_nominal.php
@@ -26,7 +26,7 @@ class AlterTableFinanceMutationSchedulesModifyNominal extends Migration
     public function down()
     {
         Schema::table("finance_mutation_schedules",function(Blueprint $table){
-            $table->decimal("nominal",20,3)->change();
+            $table->float("nominal")->change();
         });
     }
 }
